server: Fix bug in /api/event/:id
Fixes #519. There are events of the kind `taxonomy/term/associate` and
`taxonomy/term/dissociate` where the child uuid might be null (This might
be due to deleted uuids from the database). Also for some events of the
kind `taxonomy/term/parent/change` the previous or the new parent is set
to `"no parent"` in the database. This might be du to a bug or an old
behavior of the Serlo software.