Add profile images to discussions
See https://github.com/serlo/frontend/pull/966 for the PR in the new
design. Since this is just a small change and currently logged in users
only see the old design I already want to include this into the legacy
system so that authors are motivated to update their community chat
icon.